[FIX] account: compute of amount residual in foreign currency
When having an invoice in foreign currency and a partial payment in
domestic currency (in a different date with a different exchange rate),
the computed amount due is incorrect. When computing the
amount_residual_currency the payment (in domestic currency) is converted
to the foreign currency using the exchange rate of the invoice date, and
not the one of the payment date (see c0905b1fbc1ca9b1783b68fcb62f093a2e3d628f).

If the account.partial.reconcile line will have a currency_id and a
amount_currency the amount_residual_currency will not be converted but
simply added.

This behaviour was introduced with 1a6b818aa52eda37cd5b1ff26d12625262211952
and backported to version 11.0 with 3363d7235137f286458c734a1a0417beeb1f66bb.
In version 12.0 this behaviour was partially reverted in 95237ce69d5df51a78ebf09ab5f09e82fbd924d7.

opw-2156001

closes odoo/odoo#44181

X-original-commit: ed33e6dc8e3b7531ac337728e4bb99470db15580
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>