[FIX] web:tests: remove promise constructor antipattern
Revision on https://github.com/odoo/odoo/commit/6c58299874a27a4318068ee8841038da02392ee9

Having async/await as a new promise constructor is an antipattern.

The reason is that it may lead to unnoticed errors: if an inner
promise is rejected, it won't propagate the error to another promise
that will handle the error. As a result, this error becomes unnoticed
and the initial promise is pending indefinitely, which also may lead
to more bugs [1].

[1] https://stackoverflow.com/a/25569299

Closes #43072